requirements.txt
# ML & Data
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# MLflow
mlflow>=2.8.0

# Web
flask>=3.0.0
gunicorn>=21.0.0
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
